Make sure the expected Java home is always used and set for the debug service
